Ornements

+++ src/DisplayFiboIter.java
@@ -15,10 +15,10 @@ public class DisplayFiboIter {
 	 * @return
 	 */
 
-	public static String displayNumberIter(int nbIter) {
-		int valPrec1 = 1;
-		int valPrec2 = 0;
-		int value;
+	public static String displayNumberIter(long nbIter) {
+		long valPrec1 = 1;
+		long valPrec2 = 0;
+		long value;
 		String output = "";
 
 		for (int ind = 0; ind < nbIter; ind++) {
@@ -30,16 +30,23 @@ public class DisplayFiboIter {
 
 		return output;
 	}
-
-	public static String displayNumberRecurs(int nbre, int valprec1, int valprec2) {
-		int value=0;
+/**
+ * Recursive mode
+ * 
+ * @param nbre     : quantity of Fibonacci numbers
+ * @param valprec1 : 2 value before current
+ * @param valprec2 : 1 value before current
+ * @return
+ */
+	public static String displayNumberRecurs(long nbre, long valprec1, long valprec2) {
+		long value=0;
 		String output="";
-		if(nbre>0){
+		if(nbre>0){					// ending condition for recursive call
 			value=valprec1+valprec2;
 			valprec1=valprec2;
 			valprec2=value;
 			output+=" " +value+displayNumberRecurs(--nbre, valprec1, valprec2);
-					}
+		}
 		return output;
 	}
 }
